Extract mind points logic to domain and enforce coupon auth
- Move shared mind-points configs/helpers into `@mindpoint/domain/mind-points`
- Update web and Convex code to import canonical points logic from domain package
- Require authenticated identity for coupon validation and match coupon owner server-side
- Handle mobile splash screen hide errors safely
