+++ packages/domain/src/index.ts
@@ -1,6 +1,7 @@
 export * from "./cart";
 export * from "./checkout";
 export * from "./forms";
+export * from "./mind-points";
 export * from "./pricing";
 export * from "./referrals";
 export * from "./time";
